+++ ios/CHANGELOG.md
@@ -25,6 +25,7 @@ Line wrap the file at 100 chars.                                              Th
 
 ### Changed
 - Quantum-resistant tunnel setting is now on by default through the "Automatic" setting.
+- Add a checkbox that lets users include their account token in problem reports.
 
 ## [2025.8 - 2025-10-10]
 ### Fixed